+++ src/app/datagrid/datagrid-editors.demo.ts
@@ -143,14 +143,14 @@ export const LOOKUP_COLUMNS = [
       id: 'status',
       name: 'Status',
       field: 'status',
-      formatter: Formatters.Text,
+      formatter: Soho.Formatters.Text,
       sortable: false
     },
     {
       id: 'description',
       name: 'Description',
       field: 'description',
-      formatter: Formatters.Text,
+      formatter: Soho.Formatters.Text,
       sortable: false
     }
 ];
@@ -180,7 +180,7 @@ export const EDITORS_COLUMNS: any[] = [
     id: 'selectionCheckbox',
     sortable: false,
     resizable: false,
-    formatter: Formatters.SelectionCheckbox,
+    formatter: Soho.Formatters.SelectionCheckbox,
     align: 'center'
   },
 
@@ -188,7 +188,7 @@ export const EDITORS_COLUMNS: any[] = [
     id: 'rowStatus',
     sortable: false,
     resizable: false,
-    formatter: Formatters.Status,
+    formatter: Soho.Formatters.Status,
     width: 10,
     align: 'center'
   },
@@ -200,7 +200,7 @@ export const EDITORS_COLUMNS: any[] = [
     sortable: false,
     filterType: 'integer',
     width: 100,
-    formatter: Formatters.Readonly
+    formatter: Soho.Formatters.Readonly
   },
 
   {
@@ -210,8 +210,8 @@ export const EDITORS_COLUMNS: any[] = [
     sortable: false,
     filterType: 'text',
     width: 100,
-    formatter: Formatters.Lookup,
-    editor: Editors.Lookup,
+    formatter: Soho.Formatters.Lookup,
+    editor: Soho.Editors.Lookup,
     editorOptions: STATUS_LOOKUP_OPTIONS
   },
 
@@ -222,7 +222,7 @@ export const EDITORS_COLUMNS: any[] = [
     sortable: false,
     filterType: 'number',
     width: 105,
-    editor: Editors.Input
+    editor: Soho.Editors.Input
   },
 
   {
@@ -232,14 +232,14 @@ export const EDITORS_COLUMNS: any[] = [
     sortable: false,
     width: 105,
     showEmpty: true,
-    formatter: Formatters.Favorite,
-    editor: Editors.Favorite
+    formatter: Soho.Formatters.Favorite,
+    editor: Soho.Editors.Favorite
   },
 
-  //{ id: 'productName', name: 'Product Name', field: 'productName', sortable: false, filterType: 'text',    width: 150, formatter: Formatters.Hyperlink },
-  //{ id: 'activity',    name: 'Activity',     field: 'activity',    sortable: false, filterType: 'text',    width: 150, formatter: Formatters.Text, editor: Editors.Lookup, editorOptions: LOOKUP_OPTIONS },
-  //{ id: 'price',       name: 'Price',        field: 'price',       sortable: false, filterType: 'decimal', width: 125, formatter: Formatters.Decimal },
-  //{ id: 'orderDate',   name: 'Order Date',   field: 'orderDate',   sortable: false, filterType: 'date',                formatter: Formatters.Date, dateFormat: 'M/d/yyyy' }
+  //{ id: 'productName', name: 'Product Name', field: 'productName', sortable: false, filterType: 'text',    width: 150, formatter: Soho.Formatters.Hyperlink },
+  //{ id: 'activity',    name: 'Activity',     field: 'activity',    sortable: false, filterType: 'text',    width: 150, formatter: Soho.Formatters.Text, editor: Soho.Editors.Lookup, editorOptions: LOOKUP_OPTIONS },
+  //{ id: 'price',       name: 'Price',        field: 'price',       sortable: false, filterType: 'decimal', width: 125, formatter: Soho.Formatters.Decimal },
+  //{ id: 'orderDate',   name: 'Order Date',   field: 'orderDate',   sortable: false, filterType: 'date',                formatter: Soho.Formatters.Date, dateFormat: 'M/d/yyyy' }
 ];
 
 @Component({