+++ src/app/datagrid/datagrid-treegrid.demo.ts
@@ -33,11 +33,11 @@ export class DataGridTreeGridDemoComponent implements AfterContentInit, AfterVie
     if (!this._columns) {
       this._columns = [];
       /* tslint:disable */
-      this._columns.push({id: 'selectionCheckbox', sortable: false, resizable: false, width: 50, formatter: Formatters.SelectionCheckbox, align: 'center'});
-      this._columns.push({id: 'taskName', name: 'Task', field: 'taskName', expanded: 'expanded', formatter: Formatters.Tree, filterType: 'text', width: 250});
+      this._columns.push({id: 'selectionCheckbox', sortable: false, resizable: false, width: 50, formatter: Soho.Formatters.SelectionCheckbox, align: 'center'});
+      this._columns.push({id: 'taskName', name: 'Task', field: 'taskName', expanded: 'expanded', formatter: Soho.Formatters.Tree, filterType: 'text', width: 250});
       this._columns.push({id: 'id', name: 'Id', field: 'id', filterType: 'text', width: 25 });
       this._columns.push({id: 'desc', name: 'Description', field: 'desc', filterType: 'text', width: 200 });
-      this._columns.push({id: 'comments', name: 'Comments', field: 'comments', formatter: Formatters.Hyperlink, filterType: 'text', width: 60 });
+      this._columns.push({id: 'comments', name: 'Comments', field: 'comments', formatter: Soho.Formatters.Hyperlink, filterType: 'text', width: 60 });
       this._columns.push({id: 'time', name: 'Time', field: 'time', filterType: 'time', width: 60 });
       /* tslint:enable */
     }