+++ src/app/line/line.demo.ts
@@ -12,12 +12,12 @@ export class LineDemoComponent implements OnInit {
 
   public lineData = [{
     data: [
-      {name: 'Jan', value: 12, depth: 4},
-      {name: 'Feb', value: 11},
-      {name: 'Mar', value: 14},
-      {name: 'Apr', value: 10},
-      {name: 'May', value: 14},
-      {name: 'Jun', value: 8}
+      {name: 'Jan | 2018', value: 12, depth: 4},
+      {name: 'Jan | 2018', value: 11},
+      {name: 'Jan | 2018', value: 14},
+      {name: 'Jan | 2018', value: 10},
+      {name: 'Jan | 2018', value: 14},
+      {name: 'Jan | 2018', value: 8}
     ],
     name: 'Component A',
     id: '1'
@@ -47,6 +47,19 @@ export class LineDemoComponent implements OnInit {
 
   constructor() {}
 
-  ngOnInit() {}
+  public xAxis: {};
+
+  ngOnInit() {
+    this.xAxis = {
+      formatText: function (d) {
+        const text = d.split('|');
+        let markup = '';
+        text.map(function (mapText, i) {
+          markup += '<tspan x="0" dy="' + ((i + 1) * .55) + 'em">' + mapText.replace('|', ' ') + '</tspan>';
+        });
+        return markup;
+      }
+    };
+  }
 
 }